# docs/contracts/openapi.yaml
openapi: 3.1.0
info:
  title: Tai Manic Studios API
  version: 1.0.0
  description: Versioned backend contract for storefront and administration clients.
servers:
  - url: http://localhost:4000
    description: Local development
...
paths:
...
  /api/v1/checkout/quote:
    post:
      operationId: createCheckoutQuote
      tags: [checkout]
      summary: Quote the authoritative delivery, tax, and total for the current cart
      description: >
        A display quote for the current cart, an address, and a delivery method. The promotion is
        taken from the cart itself (applied via the cart endpoints). Placing the order recomputes
        everything authoritatively, so the quote is never trusted on its own. Tax is Nigerian VAT
        on the discounted goods subtotal; delivery is added after tax.
      requestBody:
        required: true
        content:
          application/json:
            schema: { $ref: '#/components/schemas/CheckoutQuoteInput' }
      responses:
        '200': { $ref: '#/components/responses/CheckoutQuote' }
        '400': { $ref: '#/components/responses/ApiError' }
        '409': { $ref: '#/components/responses/ApiError' }
...
components:
...
  securitySchemes:
    tms_session:
      type: apiKey
      in: cookie
      name: tms_session
    tms_admin_session:
      type: apiKey
      in: cookie
      name: tms_admin_session
  responses:
    ApiError:
      description: The request could not be completed.
      content:
        application/json:
          schema: { $ref: '#/components/schemas/ApiError' }
...
    CheckoutQuote:
      description: The authoritative quote for the current cart.
      content:
        application/json:
          schema: { $ref: '#/components/schemas/CheckoutQuoteResponse' }
...
  schemas:
...
    MoneyMinor:
      type: integer
      minimum: 1
      maximum: 100000000
      description: An amount in integer minor units, such as kobo. Money is never a float.
    Currency:
      type: string
      pattern: '^[A-Z]{3}$'
      example: NGN
    Money:
      type: object
      required: [amountMinor, currency]
      properties:
        amountMinor: { $ref: '#/components/schemas/MoneyMinor' }
        currency: { $ref: '#/components/schemas/Currency' }
...
    AppliedPromotion:
      type: object
      required: [code, label, discount]
      properties:
        code: { type: string }
        label: { type: string }
        discount: { $ref: '#/components/schemas/Money' }
...
    DeliveryMethodId:
      type: string
      enum: [STANDARD, EXPRESS]
    CheckoutAddressInput:
      type: object
      additionalProperties: false
      required: [state, city, line1]
      properties:
        state: { type: string, minLength: 1, maxLength: 100 }
        city: { type: string, minLength: 1, maxLength: 120 }
        line1: { type: string, minLength: 1, maxLength: 200 }
        line2: { type: string, minLength: 1, maxLength: 200, nullable: true }
        postcode: { type: string, minLength: 1, maxLength: 20, nullable: true }
...
    CheckoutQuoteInput:
      type: object
      additionalProperties: false
      required: [address, deliveryMethodId]
      properties:
        address: { $ref: '#/components/schemas/CheckoutAddressInput' }
        deliveryMethodId: { $ref: '#/components/schemas/DeliveryMethodId' }
...
    DeliveryEtaDays:
      type: object
      required: [min, max]
      properties:
        min: { type: integer, minimum: 0 }
        max: { type: integer, minimum: 0 }
    DeliveryMethod:
      type: object
      required: [id, label, description, etaDays]
      properties:
        id: { $ref: '#/components/schemas/DeliveryMethodId' }
        label: { type: string }
        description: { type: string }
        etaDays: { $ref: '#/components/schemas/DeliveryEtaDays' }
...
    CheckoutQuoteData:
      type: object
      required: [currency, subtotal, discount, delivery, tax, total, deliveryMethod, promotion]
      description: >
        Every amount is integer minor units and totals are integer arithmetic. Tax is Nigerian VAT
        on the discounted goods subtotal; the total is subtotal minus discount plus delivery plus
        tax.
      properties:
        currency: { $ref: '#/components/schemas/Currency' }
        subtotal: { $ref: '#/components/schemas/Money' }
        discount: { $ref: '#/components/schemas/Money' }
        delivery: { $ref: '#/components/schemas/Money' }
        tax: { $ref: '#/components/schemas/Money' }
        total: { $ref: '#/components/schemas/Money' }
        deliveryMethod: { $ref: '#/components/schemas/DeliveryMethod' }
        promotion:
          allOf: [{ $ref: '#/components/schemas/AppliedPromotion' }]
          nullable: true
    CheckoutQuoteResponse:
      type: object
      required: [data, meta]
      properties:
        data: { $ref: '#/components/schemas/CheckoutQuoteData' }
        meta: { $ref: '#/components/schemas/ApiMeta' }
...
    ApiMeta:
      type: object
      required: [correlationId]
      properties:
        correlationId: { type: string }
    ApiError:
      type: object
      required: [error]
      properties:
        error:
          type: object
          required: [code, message, correlationId]
          properties:
            code:
              type: string
              enum:
                [
                  VALIDATION_FAILED,
                  AUTHENTICATION_REQUIRED,
                  AUTHENTICATION_INVALID,
                  EMAIL_VERIFICATION_REQUIRED,
                  TOKEN_INVALID_OR_EXPIRED,
                  SESSION_INVALID,
                  ADMIN_MFA_REQUIRED,
                  MFA_CHALLENGE_INVALID,
                  MFA_CODE_INVALID,
                  PERMISSION_DENIED,
                  RESOURCE_NOT_FOUND,
                  CONFLICT,
                  RATE_LIMITED,
                  IDEMPOTENCY_CONFLICT,
                  INVENTORY_UNAVAILABLE,
                  PROMOTION_INVALID,
                  CONFIGURATION_NOT_APPROVED,
                  MEDIA_VALIDATION_FAILED,
                  MEDIA_INFECTED,
                  MEDIA_PROCESSING_FAILED,
                  INTEGRATION_UNAVAILABLE,
                  INTERNAL_ERROR,
                ]
            message: { type: string }
            correlationId: { type: string }